requirements.txt
opencv-python
pillow
numpy
pygame
scikit-image